small-case-study/alfa/requirements.txt
Flask
Jinja2
kafka-python
Faker